requirements.txt
python-alipay-sdk==3.0.4
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
Deprecated==1.2.13
Django==3.2.13
django-haystack==3.2.1
django-redis==5.2.0
django-tinymce==3.4.0
fdfs-client-py @ file:///home/user/Desktop/fdfs_client-py-master.zip
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
jieba==0.42.1
kombu==5.1.0
mutagen==1.45.1
packaging==21.3
Pillow==8.4.0
pip==21.3.1
prompt-toolkit==3.0.29
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.9.4
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
rsa==4.8
setuptools==59.6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
Whoosh==2.7.4
wrapt==1.14.0
zipp==3.6.0
